Refactor: replaced inheritance with composition in  metadata field validators
