+++ src/datasets/domain/useCases/validators/MetadataFieldValidator.ts
@@ -1,13 +1,17 @@
-import {
-  BaseMetadataFieldValidator,
-  NewDatasetMetadataFieldAndValueInfo,
-} from './BaseMetadataFieldValidator';
+import { BaseMetadataFieldValidator, NewDatasetMetadataFieldAndValueInfo } from './BaseMetadataFieldValidator';
 import { MultipleMetadataFieldValidator } from './MultipleMetadataFieldValidator';
 import { SingleMetadataFieldValidator } from './SingleMetadataFieldValidator';
 import { EmptyFieldError } from './errors/EmptyFieldError';
 import { NewDatasetMetadataFieldValueDTO } from '../../dtos/NewDatasetDTO';
 
 export class MetadataFieldValidator extends BaseMetadataFieldValidator {
+  constructor(
+    private singleMetadataFieldValidator: SingleMetadataFieldValidator,
+    private multipleMetadataFieldValidator: MultipleMetadataFieldValidator,
+  ) {
+    super();
+  }
+
   validate(newDatasetMetadataFieldAndValueInfo: NewDatasetMetadataFieldAndValueInfo): void {
     const metadataFieldValue = newDatasetMetadataFieldAndValueInfo.metadataFieldValue;
     const metadataFieldInfo = newDatasetMetadataFieldAndValueInfo.metadataFieldInfo;
@@ -29,12 +33,9 @@ export class MetadataFieldValidator extends BaseMetadataFieldValidator {
       }
     }
     if (newDatasetMetadataFieldAndValueInfo.metadataFieldInfo.multiple) {
-      this.executeMetadataFieldValidator(
-        new MultipleMetadataFieldValidator(),
-        newDatasetMetadataFieldAndValueInfo,
-      );
+      this.multipleMetadataFieldValidator.validate(newDatasetMetadataFieldAndValueInfo);
     } else {
-      this.executeMetadataFieldValidator(new SingleMetadataFieldValidator(), newDatasetMetadataFieldAndValueInfo);
+      this.singleMetadataFieldValidator.validate(newDatasetMetadataFieldAndValueInfo);
     }
   }
