+++ src/datasets/domain/useCases/validators/SingleMetadataFieldValidator.ts
@@ -1,7 +1,9 @@
-import {
-  BaseMetadataFieldValidator,
-  NewDatasetMetadataFieldAndValueInfo,
-} from './BaseMetadataFieldValidator';
+import { BaseMetadataFieldValidator, NewDatasetMetadataFieldAndValueInfo } from './BaseMetadataFieldValidator';
+import { ControlledVocabularyFieldError } from './errors/ControlledVocabularyFieldError';
+import { DateFormatFieldError } from './errors/DateFormatFieldError';
+import { MetadataFieldValidator } from './MetadataFieldValidator';
+import { NewDatasetMetadataChildFieldValueDTO } from '../../dtos/NewDatasetDTO';
+import { MultipleMetadataFieldValidator } from './MultipleMetadataFieldValidator';
 
 export class SingleMetadataFieldValidator extends BaseMetadataFieldValidator {
   validate(newDatasetMetadataFieldAndValueInfo: NewDatasetMetadataFieldAndValueInfo): void {
@@ -26,5 +28,68 @@ export class SingleMetadataFieldValidator extends BaseMetadataFieldValidator {
       );
     }
     this.validateFieldValue(newDatasetMetadataFieldAndValueInfo);
+  }
+
+  private validateFieldValue(newDatasetMetadataFieldAndValueInfo: NewDatasetMetadataFieldAndValueInfo) {
+    const metadataFieldInfo = newDatasetMetadataFieldAndValueInfo.metadataFieldInfo;
+    if (metadataFieldInfo.isControlledVocabulary) {
+      this.validateControlledVocabularyFieldValue(newDatasetMetadataFieldAndValueInfo);
+    }
+
+    if (metadataFieldInfo.type == 'DATE') {
+      this.validateDateFieldValue(newDatasetMetadataFieldAndValueInfo);
+    }
+
+    if (metadataFieldInfo.childMetadataFields != undefined) {
+      this.validateChildMetadataFieldValues(newDatasetMetadataFieldAndValueInfo);
+    }
+  }
+
+  private validateControlledVocabularyFieldValue(
+    newDatasetMetadataFieldAndValueInfo: NewDatasetMetadataFieldAndValueInfo,
+  ) {
+    if (
+      !newDatasetMetadataFieldAndValueInfo.metadataFieldInfo.controlledVocabularyValues.includes(
+        newDatasetMetadataFieldAndValueInfo.metadataFieldValue as string,
+      )
+    ) {
+      throw new ControlledVocabularyFieldError(
+        newDatasetMetadataFieldAndValueInfo.metadataFieldKey,
+        newDatasetMetadataFieldAndValueInfo.metadataBlockName,
+        newDatasetMetadataFieldAndValueInfo.metadataParentFieldKey,
+        newDatasetMetadataFieldAndValueInfo.metadataFieldPosition,
+      );
+    }
+  }
+
+  private validateDateFieldValue(newDatasetMetadataFieldAndValueInfo: NewDatasetMetadataFieldAndValueInfo) {
+    const dateFormatRegex = /^\d{4}-\d{2}-\d{2}$/;
+    if (!dateFormatRegex.test(newDatasetMetadataFieldAndValueInfo.metadataFieldValue as string)) {
+      throw new DateFormatFieldError(
+        newDatasetMetadataFieldAndValueInfo.metadataFieldKey,
+        newDatasetMetadataFieldAndValueInfo.metadataBlockName,
+        newDatasetMetadataFieldAndValueInfo.metadataParentFieldKey,
+        newDatasetMetadataFieldAndValueInfo.metadataFieldPosition,
+      );
+    }
+  }
+
+  private validateChildMetadataFieldValues(newDatasetMetadataFieldAndValueInfo: NewDatasetMetadataFieldAndValueInfo) {
+    const metadataFieldInfo = newDatasetMetadataFieldAndValueInfo.metadataFieldInfo;
+    const childMetadataFieldKeys = Object.keys(metadataFieldInfo.childMetadataFields);
+    const metadataFieldValidator = new MetadataFieldValidator(this, new MultipleMetadataFieldValidator(this));
+    for (const childMetadataFieldKey of childMetadataFieldKeys) {
+      const childMetadataFieldInfo = metadataFieldInfo.childMetadataFields[childMetadataFieldKey];
+      metadataFieldValidator.validate({
+        metadataFieldInfo: childMetadataFieldInfo,
+        metadataFieldKey: childMetadataFieldKey,
+        metadataFieldValue: (
+          newDatasetMetadataFieldAndValueInfo.metadataFieldValue as NewDatasetMetadataChildFieldValueDTO
+        )[childMetadataFieldKey],
+        metadataBlockName: newDatasetMetadataFieldAndValueInfo.metadataBlockName,
+        metadataParentFieldKey: newDatasetMetadataFieldAndValueInfo.metadataFieldKey,
+        metadataFieldPosition: newDatasetMetadataFieldAndValueInfo.metadataFieldPosition,
+      });
+    }
   }
 }
